+++ test/support/response_code.ex
@@ -2,6 +2,7 @@ defmodule TdDdWeb.ResponseCode do
   @moduledoc false
 
   @ok "Ok"
+  @accepted "Accepted"
   @created "Created"
   @no_content "No Content"
   @forbidden "Forbidden"
@@ -10,16 +11,19 @@ defmodule TdDdWeb.ResponseCode do
   @unprocessable_entity "Unprocessable Entity"
 
   def rc_ok, do: @ok
+  def rc_accepted, do: @accepted
   def rc_created, do: @created
   def rc_forbidden, do: @forbidden
   def rc_unauthorized, do: @unauthorized
   def rc_not_found, do: @not_found
   def rc_unprocessable_entity, do: @unprocessable_entity
 
+  # credo:disable-for-next-line Credo.Check.Refactor.CyclomaticComplexity
   def to_response_code(http_status_code) do
     case http_status_code do
       200 -> @ok
       201 -> @created
+      202 -> @accepted
       204 -> @no_content
       401 -> @unauthorized
       403 -> @forbidden